+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 - Show icon next to spell/item ID
 - Show spell/item tooltip when hovering over the ID
 - Add item equipped load option to bindings
+- Add tooltips to load options
 
 ### Fixed